Refactored React AntD scaffolding and applied the Adapter pattern for multi-UI page generation
