+++ CHANGELOG.md
@@ -86,3 +86,6 @@
 
 ## 1.0.25 — 2025-09-25
 - Performed refactoring on React Tailwind scaffolding pages
+
+## 1.0.26 — 2025-09-26
+- Refactored React AntD scaffolding and applied the Adapter pattern for multi-UI page generation
