+++ README.md
@@ -1,4 +1,4 @@
-# polyfront-scaffold (v1.0.25)
+# polyfront-scaffold (v1.0.26)
 
 Author: **Nirmal Samaranayaka** <nirmal.fullstack@gmail.com>  
 Repo: https://github.com/NirmalSamaranayaka/polyfront-scaffold
