+++ lib/pages/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  ...require("./adapters/mui"),
+  ...require("./adapters/bootstrap"),
+  ...require("./adapters/antd"),
+  ...require("./adapters/tailwind"),
+};
