+++ lib/scaffold-react-project-structure.js
@@ -1,9 +1,7 @@
 // lib/scaffold-react-project-structure.js
 const fs = require("fs");
 const path = require("path");
-const { createMUIPages } = require("./pages/mui");
-const { createBootstrapPages } = require("./pages/bootstrap");
-const { createTailwindPages } = require("./pages/tailwind");
+const { createMUIPages, createBootstrapPages, createTailwindPages, createAntdPages } = require("./pages/index");
 
 function writeFileSafe(p, content) {
   fs.mkdirSync(path.dirname(p), { recursive: true });
@@ -141,550 +139,6 @@ export const router = createBrowserRouter([
     // Default to MUI if no specific UI is chosen
     createMUIPages(pagesDir, extx, ext, useTS);
   }
-}
-
-function createAntdPages(pagesDir, extx, ext, useTS) {
-   const home = path.join(pagesDir, `Home.${extx}`);
-  if (!fs.existsSync(home)) {
-    writeFileSafe(home, `import { Typography, Row, Col, Card, Button, Avatar, theme } from 'antd';
-import {
-  RiseOutlined,
-  TeamOutlined,
-  ThunderboltOutlined,
-  SafetyOutlined,
-  CodeOutlined,
-  BgColorsOutlined,
-} from '@ant-design/icons';
-
-const { Title, Paragraph, Text } = Typography;
-
-export default function Home() {
-  const { token } = theme.useToken();
-
-  const colorMap${useTS?': Record<string, string>':''} = {
-    primary: token.colorPrimary,
-    secondary: token.colorInfo,
-    success: token.colorSuccess,
-    warning: token.colorWarning,
-    info: token.colorInfo,
-    error: token.colorError,
-  };
-
-  const features = [
-    { icon: <RiseOutlined />, title: 'Performance', description: 'Lightning-fast builds with modern tooling', color: 'primary' },
-    { icon: <TeamOutlined />, title: 'Team Ready', description: 'Built for collaboration and scalability', color: 'secondary' },
-    { icon: <ThunderboltOutlined />, title: 'Fast Development', description: 'Hot reload and instant feedback', color: 'success' },
-    { icon: <SafetyOutlined />, title: 'Type Safe', description: 'Full TypeScript support with strict mode', color: 'warning' },
-    { icon: <CodeOutlined />, title: 'Clean Code', description: 'ESLint + Prettier for code quality', color: 'info' },
-    { icon: <BgColorsOutlined />, title: 'Beautiful UI', description: 'Ant Design components with theming', color: 'error' },
-  ];
-
-  return (
-    <div style={{ maxWidth: 1200, margin: '0 auto', padding: 32 }}>
-      <div style={{ textAlign: 'center', marginBottom: 48 }}>
-        <Title level={1} style={{ color: token.colorPrimary, fontWeight: 800, marginBottom: 8 }}>
-          Welcome to PolyFront
-        </Title>
-        <Title level={4} type="secondary" style={{ marginBottom: 12 }}>
-          Your comprehensive multi-frontend scaffold with everything you need
-        </Title>
-        <Paragraph type="secondary" style={{ maxWidth: 600, margin: '0 auto' }}>
-          Built with React, TypeScript, Vite, Ant Design, and modern development practices.
-          Start building your next great application today.
-        </Paragraph>
-      </div>
-
-      <Row gutter={[24, 24]} style={{ marginBottom: 48 }}>
-        {features.map((feature, idx) => (
-          <Col key={idx} xs={24} sm={12} md={8}>
-            <Card
-              hoverable
-              style={{ height: '100%', transition: 'transform .2s' }}
-              bodyStyle={{ textAlign: 'center' }}
-              onMouseEnter={(e) => (e.currentTarget.style.transform = 'translateY(-4px)')}
-              onMouseLeave={(e) => (e.currentTarget.style.transform = 'translateY(0)')}
-            >
-              <Avatar
-                size={64}
-                style={{
-                  backgroundColor: colorMap[feature.color] ?? token.colorPrimary,
-                  margin: '0 auto 12px',
-                  display: 'grid',
-                  placeItems: 'center',
-                }}
-              >
-                {feature.icon}
-              </Avatar>
-              <Title level={5} style={{ marginBottom: 8 }}>{feature.title}</Title>
-              <Text type="secondary">{feature.description}</Text>
-            </Card>
-          </Col>
-        ))}
-      </Row>
-
-      <Card style={{ textAlign: 'center' }}>
-        <Title level={2} style={{ marginBottom: 8 }}>Ready to Get Started?</Title>
-        <Paragraph type="secondary" style={{ maxWidth: 760, margin: '0 auto 16px' }}>
-          This scaffold includes routing, state management, API integration, testing, and much more.
-          Explore the different pages to see all the features in action.
-        </Paragraph>
-        <div style={{ marginTop: 16 }}>
-          <Button type="primary" size="large" style={{ marginRight: 12 }}>
-            View Documentation
-          </Button>
-          <Button size="large">Learn More</Button>
-        </div>
-      </Card>
-    </div>
-  );
-}
-`);
-  }
-
-  // Create other Antd pages
-  createOtherAntdPages(pagesDir, extx, ext, useTS);
-}
-
-function createOtherAntdPages(pagesDir, extx, ext, useTS) {
-   // About page
-  const about = path.join(pagesDir, `About.${extx}`);
-  if (!fs.existsSync(about)) {
-    writeFileSafe(about, `import { Typography, Row, Col, Card, List, Divider, Tag } from 'antd';
-import { CheckCircleTwoTone, StarFilled, ThunderboltOutlined, TeamOutlined } from '@ant-design/icons';
-
-export default function About() {
-  const techStack = [
-    { name: 'React 19', version: 'Latest', category: 'Framework' },
-    { name: 'TypeScript', version: '5.x', category: 'Language' },
-    { name: 'Vite', version: '7.x', category: 'Build Tool' },
-    { name: 'Ant Design', version: '5.x', category: 'UI Library' },
-    { name: 'React Router', version: '7.x', category: 'Routing' },
-    { name: 'Zustand', version: '5.x', category: 'State Management' },
-    { name: 'Axios', version: '1.x', category: 'HTTP Client' },
-    { name: 'Vitest', version: '3.x', category: 'Testing' }
-  ];
-
-  const features = [
-    'Modern React with hooks and functional components',
-    'Full TypeScript support with strict mode',
-    'Beautiful Ant Design components with theming',
-    'Client-side routing with React Router',
-    'State management with Zustand',
-    'API integration with Axios',
-    'Comprehensive testing setup with Vitest',
-    'ESLint and Prettier for code quality',
-    'Hot module replacement for fast development',
-    'Responsive design with mobile-first approach'
-  ];
-
-  return (
-    <div style={{ maxWidth: 1200, margin: '0 auto', padding: 24 }}>
-      <div style={{ textAlign: 'center', marginBottom: 24 }}>
-        <Typography.Title level={1} style={{ fontWeight: 800 }}>
-          About PolyFront
-        </Typography.Title>
-        <Typography.Paragraph type="secondary">
-          A comprehensive scaffold for modern web applications
-        </Typography.Paragraph>
-      </div>
-
-      <Row gutter={[24, 24]} style={{ marginBottom: 24 }}>
-        <Col xs={24} md={12}>
-          <Card title={<Typography.Title level={3} style={{ margin: 0 }}>Technology Stack</Typography.Title>}>
-            <Typography.Paragraph>
-              Built with the latest and greatest technologies in the React ecosystem.
-            </Typography.Paragraph>
-            <Row gutter={[12, 12]}>
-              {techStack.map((tech, idx) => (
-                <Col xs={12} key={idx}>
-                  <Tag style={{ marginBottom: 4 }}>
-                    {\`\${tech.name} \${tech.version}\`}
-                  </Tag>
-                  <div style={{ color: '#999', fontSize: 12 }}>{tech.category}</div>
-                </Col>
-              ))}
-            </Row>
-          </Card>
-        </Col>
-
-        <Col xs={24} md={12}>
-          <Card title={<Typography.Title level={3} style={{ margin: 0 }}>Key Features</Typography.Title>}>
-            <List
-              itemLayout="horizontal"
-              dataSource={features}
-              renderItem={(text, index) => (
-                <>
-                  <List.Item>
-                    <List.Item.Meta
-                      avatar={<CheckCircleTwoTone twoToneColor="#52c41a" />}
-                      title={<span>{text}</span>}
-                    />
-                  </List.Item>
-                  {index < features.length - 1 && <Divider style={{ margin: 0 }} />}
-                </>
-              )}
-            />
-          </Card>
-        </Col>
-      </Row>
-
-      <Card>
-        <Row gutter={[24, 24]} align="middle" justify="center">
-          <Col xs={24} sm={8} style={{ textAlign: 'center' }}>
-            <StarFilled style={{ fontSize: 40, color: '#1677ff', marginBottom: 8 }} />
-            <Typography.Title level={4}>Production Ready</Typography.Title>
-            <Typography.Paragraph type="secondary">Built with best practices and industry standards</Typography.Paragraph>
-          </Col>
-          <Col xs={24} sm={8} style={{ textAlign: 'center' }}>
-            <ThunderboltOutlined style={{ fontSize: 40, color: '#1677ff', marginBottom: 8 }} />
-            <Typography.Title level={4}>Fast Development</Typography.Title>
-            <Typography.Paragraph type="secondary">Hot reload and instant feedback for rapid iteration</Typography.Paragraph>
-          </Col>
-          <Col xs={24} sm={8} style={{ textAlign: 'center' }}>
-            <TeamOutlined style={{ fontSize: 40, color: '#1677ff', marginBottom: 8 }} />
-            <Typography.Title level={4}>Team Friendly</Typography.Title>
-            <Typography.Paragraph type="secondary">Consistent code style and comprehensive tooling</Typography.Paragraph>
-          </Col>
-        </Row>
-      </Card>
-    </div>
-  );
-}
-`);
-  }
-
-  // Dashboard page
-  const dashboard = path.join(pagesDir, `Dashboard.${extx}`);
-  if (!fs.existsSync(dashboard)) {
-      writeFileSafe(dashboard, `import React, { useState } from 'react';
-import {
-  Typography,
-  Row,
-  Col,
-  Card,
-  Avatar,
-  Table,
-  Tag,
-  List,
-  theme,
-  Button,
-} from 'antd';
-import {
-  UserOutlined,
-  DollarOutlined,
-  ShoppingCartOutlined,
-  EyeOutlined,
-  ArrowUpOutlined,
-  ArrowDownOutlined,
-  CheckCircleTwoTone,
-  ExclamationCircleTwoTone,
-  CloseCircleTwoTone,
-} from '@ant-design/icons';
-
-export default function Dashboard() {
-  const { token } = theme.useToken();
-
-  const [stats] = useState({
-    users: 1234,
-    revenue: 45678,
-    orders: 89,
-    views: 12345,
-  });
-
-  const [recentActivity] = useState([
-    { id: 1, user: 'Nirmal Samaranayaka', action: 'Order placed', time: '2 min ago', status: 'success' },
-    { id: 2, user: 'Jane Smith', action: 'Payment received', time: '5 min ago', status: 'success' },
-    { id: 3, user: 'Bob Johnson', action: 'Login attempt', time: '10 min ago', status: 'warning' },
-    { id: 4, user: 'Alice Brown', action: 'Account created', time: '15 min ago', status: 'success' },
-  ]);
-
-  const [topProducts] = useState([
-    { key: '1', name: 'Premium Widget', sales: 156, revenue: 2340, growth: 12 },
-    { key: '2', name: 'Super Gadget', sales: 98, revenue: 1890, growth: -5 },
-    { key: '3', name: 'Mega Tool', sales: 87, revenue: 1450, growth: 8 },
-    { key: '4', name: 'Ultra Device', sales: 76, revenue: 1230, growth: 15 },
-  ]);
-
-  const growthTag = (g ${useTS?': number':''}) => (
-    <Tag color={g >= 0 ? 'success' : 'error'}>
-      {g >= 0 ? <ArrowUpOutlined /> : <ArrowDownOutlined />} {g}%
-    </Tag>
-  );
-
-  const columns = [
-    { title: 'Product', dataIndex: 'name' },
-    { title: 'Sales', dataIndex: 'sales', align: 'right' ${useTS?' as const':''} },
-    { title: 'Revenue', dataIndex: 'revenue', align: 'right' ${useTS?' as const':''}, render: (v ${useTS?': number':''}) => \`$\${v.toLocaleString()}\` },
-    { title: 'Growth', dataIndex: 'growth', align: 'right'${useTS?' as const':''} , render: (g${useTS?': number':''}) => growthTag(g) },
-  ];
-
-  const StatCard ${useTS?': React.FC<{ title: string; value: React.ReactNode; sub: string; avatarBg: string; icon: React.ReactNode }>':''} 
-    = ({ title, value, sub, avatarBg, icon }) => (
-      <Card>
-        <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between' }}>
-          <div>
-            <Typography.Text type="secondary">{title}</Typography.Text>
-            <Typography.Title level={3} style={{ margin: 0 }}>{value}</Typography.Title>
-            <Typography.Text style={{ color: token.colorSuccess }}>
-              <ArrowUpOutlined style={{ marginRight: 6 }} />{sub}
-            </Typography.Text>
-          </div>
-          <Avatar size={56} style={{ background: avatarBg, color: '#fff' }}>
-            {icon}
-          </Avatar>
-        </div>
-      </Card>
-    );
-
-  const statusIcon = (status${useTS?': string':''}) => {
-    switch (status) {
-      case 'success':
-        return <CheckCircleTwoTone twoToneColor="#52c41a" />;
-      case 'warning':
-        return <ExclamationCircleTwoTone twoToneColor="#faad14" />;
-      case 'error':
-        return <CloseCircleTwoTone twoToneColor="#ff4d4f" />;
-      default:
-        return <CheckCircleTwoTone twoToneColor="#52c41a" />;
-    }
-  };
-
-  return (
-    <div>
-      <Typography.Title level={1} style={{ color: token.colorPrimary, fontWeight: 800 }}>Dashboard</Typography.Title>
-      <Typography.Paragraph type="secondary">
-        Welcome back! Here's what's happening with your application today.
-      </Typography.Paragraph>
-
-      {/* Stats */}
-      <Row gutter={[16, 16]} style={{ marginBottom: 16 }}>
-        <Col xs={24} sm={12} md={6}>
-          <StatCard title="Total Users" value={stats.users.toLocaleString()} sub="+12% from last month" avatarBg={token.colorPrimary} icon={<UserOutlined />} />
-        </Col>
-        <Col xs={24} sm={12} md={6}>
-          <StatCard title="Revenue" value={\`$\${stats.revenue.toLocaleString()}\`} sub="+8% from last month" avatarBg={token.colorSuccess} icon={<DollarOutlined />} />
-        </Col>
-        <Col xs={24} sm={12} md={6}>
-          <StatCard title="Orders" value={stats.orders} sub="+15% from last month" avatarBg={token.colorWarning} icon={<ShoppingCartOutlined />} />
-        </Col>
-        <Col xs={24} sm={12} md={6}>
-          <StatCard title="Page Views" value={stats.views.toLocaleString()} sub="+22% from last month" avatarBg={token.colorInfo} icon={<EyeOutlined />} />
-        </Col>
-      </Row>
-
-      <Row gutter={[16, 16]}>
-        <Col xs={24} md={16}>
-          <Card title="Top Products" extra={<Button type="link">View All</Button>}>
-            <Table columns={columns} dataSource={topProducts} pagination={false} />
-          </Card>
-        </Col>
-        <Col xs={24} md={8}>
-          <Card title="Recent Activity">
-            <List
-              dataSource={recentActivity}
-              renderItem={(item) => (
-                <List.Item>
-                  <List.Item.Meta
-                    avatar={<Avatar style={{ background: '#f5f5f5' }}>{statusIcon(item.status)}</Avatar>}
-                    title={item.action}
-                    description={<span><strong>{item.user}</strong> — {item.time}</span>}
-                  />
-                </List.Item>
-              )}
-            />
-          </Card>
-        </Col>
-      </Row>
-    </div>
-  );
-}`);
-  }
-
-  // Profile page
-    const profile = path.join(pagesDir, `Profile.${extx}`);
-    if (!fs.existsSync(profile)) {
-      writeFileSafe(profile, `import { useState } from 'react';
- import {
-   Typography,
-   Row,
-   Col,
-   Card,
-   Avatar,
-   Button,
-   Tag,
-   List,
-   Switch,
-   Form,
-   Input,
- } from 'antd';
- import {
-   EditOutlined,
-   SaveOutlined,
-   CloseOutlined,
-   MailOutlined,
-   PhoneOutlined,
-   EnvironmentOutlined,
-   BankOutlined,
-   BellOutlined,
-   SafetyOutlined,
-   BgColorsOutlined,
-   UserOutlined,
- } from '@ant-design/icons';
- 
- export default function Profile() {
-   const [isEditing, setIsEditing] = useState(false);
-   const [settings, setSettings] = useState({
-     emailNotifications: true,
-     pushNotifications: false,
-     darkMode: false,
-     twoFactorAuth: true,
-   });
- 
-   const profile = {
-     name: 'Nirmal Samaranayaka',
-     email: 'nirmal.fullstack@gmail.com',
-     phone: '+46 (72) xxx-xxxx',
-     location: 'Stockholm, Sweden',
-     company: 'Scania AB.',
-     position: 'Senior Fullstack Developer',
-     education: 'Computer Science, University of Colombo',
-     website: 'https://dev.to/nirmalsamaranayaka',
-     bio: 'Experienced Full Stack Engineer & Tech Lead | Specialized in .NET, React, Angular, and scalable cloud-native solutions.',
-   };
- 
-   const skills = [
-     'React', 'Redux', 'MobX', 'React Query', 'Angular', 'Micro Frontends',
-     'TypeScript', 'JavaScript', 'HTML/CSS', 'jQuery',
-     'MUI (Material-UI)', 'Kendo UI', 'Tailwind CSS', 'Tegel',
-     '.NET 6/7/8', 'ASP.NET Core', 'C#', 'Web API', 'WCF',
-     'Microservices', 'gRPC', 'REST APIs', 'SOAP', 'Node.js',
-     'Solidity', 'Truffle',
-     'Azure', 'AWS', 'Docker', 'Kubernetes', 'Helm',
-     'CI/CD', 'Azure DevOps', 'Jenkins', 'Git', 'TFS/TFVC',
-     'MSSQL', 'Oracle', 'Entity Framework Core', 'Entity Framework 6', 'Dapper',
-     'T-SQL', 'PL-SQL', 'SSIS', 'SSRS', 'Data Migration', 'Query Tuning',
-     'Clean Architecture', 'Domain-Driven Design (DDD)', 'SOLID',
-     'OOP', 'TDD', 'MVVM', 'Agile (Scrum, Kanban)'
-   ];
- 
-   const onSettingChange = (key ${useTS?': keyof typeof settings':''}) => (checked ${useTS?': boolean':''}) =>
-     setSettings(prev => ({ ...prev, [key]: checked }));
- 
-   return (
-     <div>
-       <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: 16 }}>
-         <Typography.Title level={1} style={{ margin: 0 }}>Profile</Typography.Title>
-         <div>
-           {isEditing ? (
-             <>
-               <Button type="primary" icon={<SaveOutlined />} onClick={() => setIsEditing(false)} style={{ marginRight: 8 }}>Save</Button>
-               <Button icon={<CloseOutlined />} onClick={() => setIsEditing(false)}>Cancel</Button>
-             </>
-           ) : (
-             <Button type="primary" icon={<EditOutlined />} onClick={() => setIsEditing(true)}>Edit Profile</Button>
-           )}
-         </div>
-       </div>
- 
-       <Row gutter={[16, 16]}>
-         {/* Header */}
-         <Col span={24}>
-           <Card>
-             <div style={{ display: 'flex', gap: 16, alignItems: 'center' }}>
-               <Avatar size={120} icon={<UserOutlined />} />
-               <div style={{ flex: 1 }}>
-                 <Typography.Title level={3} style={{ marginBottom: 4 }}>{profile.name}</Typography.Title>
-                 <Typography.Text type="secondary" style={{ display: 'block', marginBottom: 8 }}>
-                   {profile.position} at {profile.company}
-                 </Typography.Text>
-                 <Typography.Paragraph type="secondary" style={{ marginBottom: 12 }}>{profile.bio}</Typography.Paragraph>
-                 <div style={{ display: 'flex', gap: 8, flexWrap: 'wrap' }}>
-                   {skills.map((s) => (
-                     <Tag key={s}>{s}</Tag>
-                   ))}
-                 </div>
-               </div>
-             </div>
-           </Card>
-         </Col>
- 
-         {/* Info */}
-         <Col xs={24} md={16}>
-           <Card title="Personal Information">
-             <Form layout="vertical" disabled={!isEditing}>
-               <Row gutter={12}>
-                 <Col xs={24} sm={12}>
-                   <Form.Item label="Full Name">
-                     <Input defaultValue={profile.name} prefix={<MailOutlined />} />
-                   </Form.Item>
-                 </Col>
-                 <Col xs={24} sm={12}>
-                   <Form.Item label="Email">
-                     <Input defaultValue={profile.email} prefix={<MailOutlined />} />
-                   </Form.Item>
-                 </Col>
-                 <Col xs={24} sm={12}>
-                   <Form.Item label="Phone">
-                     <Input defaultValue={profile.phone} prefix={<PhoneOutlined />} />
-                   </Form.Item>
-                 </Col>
-                 <Col xs={24} sm={12}>
-                   <Form.Item label="Location">
-                     <Input defaultValue={profile.location} prefix={<EnvironmentOutlined />} />
-                   </Form.Item>
-                 </Col>
-                 <Col xs={24} sm={12}>
-                   <Form.Item label="Company">
-                     <Input defaultValue={profile.company} prefix={<BankOutlined />} />
-                   </Form.Item>
-                 </Col>
-                 <Col xs={24} sm={12}>
-                   <Form.Item label="Position">
-                     <Input defaultValue={profile.position} prefix={<BankOutlined />} />
-                   </Form.Item>
-                 </Col>
-                 <Col span={24}>
-                   <Form.Item label="Bio">
-                     <Input.TextArea rows={3} defaultValue={profile.bio} />
-                   </Form.Item>
-                 </Col>
-               </Row>
-             </Form>
-           </Card>
-         </Col>
- 
-         {/* Settings & Actions */}
-         <Col xs={24} md={8}>
-           <Card title="Settings" style={{ marginBottom: 16 }}>
-             <List>
-               <List.Item actions={[<Switch checked={settings.emailNotifications} onChange={onSettingChange('emailNotifications')} />]}>
-                 <List.Item.Meta avatar={<BellOutlined />} title="Email Notifications" />
-               </List.Item>
-               <List.Item actions={[<Switch checked={settings.pushNotifications} onChange={onSettingChange('pushNotifications')} />]}>
-                 <List.Item.Meta avatar={<BellOutlined />} title="Push Notifications" />
-               </List.Item>
-               <List.Item actions={[<Switch checked={settings.darkMode} onChange={onSettingChange('darkMode')} />]}>
-                 <List.Item.Meta avatar={<BgColorsOutlined />} title="Dark Mode" />
-               </List.Item>
-               <List.Item actions={[<Switch checked={settings.twoFactorAuth} onChange={onSettingChange('twoFactorAuth')} />]}>
-                 <List.Item.Meta avatar={<SafetyOutlined />} title="Two-Factor Auth" />
-               </List.Item>
-             </List>
-           </Card>
- 
-           <Card title="Quick Actions">
-             <Button block style={{ marginBottom: 8 }}>Download Resume</Button>
-             <Button block style={{ marginBottom: 8 }}>Share Profile</Button>
-             <Button block danger>Delete Account</Button>
-           </Card>
-         </Col>
-       </Row>
-     </div>
-   );
- }
- `);
-    }
-
 }
 
 function createChakraPages(pagesDir, extx, ext, useTS) {
@@ -2400,7 +1854,7 @@ const Layout = () => {
           </Typography>
 
           <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-            <Chip label="v1.0.25" size="small" variant="outlined" />
+            <Chip label="v1.0.26" size="small" variant="outlined" />
             {/* Language selector using i18n */}
             <Select
               value={currentLang}
@@ -2545,7 +1999,7 @@ const Layout = () => {
             </button>
             <h4 className="mb-0">PolyFront</h4>
             <div className="d-flex align-items-center gap-2">
-              <span className="badge bg-secondary">v1.0.25</span>
+              <span className="badge bg-secondary">v1.0.26</span>
               
               {/* Language Selector */}
               <select
@@ -2639,7 +2093,7 @@ export default Layout;`);
                 {/* Version + Language + User */}
                 <div className="flex items-center gap-3">
                   {/* Version */}
-                  <span className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm">v1.0.25</span>
+                  <span className="bg-gray-100 text-gray-700 px-2 py-1 rounded-full text-sm">v1.0.26</span>
       
                   {/* Language selector */}
                   <select
@@ -2748,7 +2202,7 @@ style={{ borderRight: 0 }}
         >
           PolyFront
         </Typography.Title>
-        <Tag bordered={false}>v1.0.25</Tag>
+        <Tag bordered={false}>v1.0.26</Tag>
         <Select
           size="small"
           value={currentLang}
@@ -2942,7 +2396,7 @@ function Layout() {
           <Box flex="1" />
 
           <HStack gap={2}>
-            <Badge>v1.0.25</Badge>
+            <Badge>v1.0.26</Badge>
 
             {/* Language Select (v3 controlled) */}
             <Select.Root
